// shared/pokemonData/models/PokemonInstance.ts
export enum Nature {
  Hardy = 0,
  Lonely = 1,
  Brave = 2,
  Adamant = 3,
  Naughty = 4,
  Bold = 5,
  Docile = 6,
  Relaxed = 7,
  Impish = 8,
  Lax = 9,
  Timid = 10,
  Hasty = 11,
  Serious = 12,
  Jolly = 13,
  Naive = 14,
  Modest = 15,
  Mild = 16,
  Quiet = 17,
  Bashful = 18,
  Rash = 19,
  Calm = 20,
  Gentle = 21,
  Sassy = 22,
  Careful = 23,
  Quirky = 24,
}

export enum Stat {
  HP = 0,
  Attack = 1,
  Defense = 2,
  SpAttack = 3,
  SpDefense = 4,
  Speed = 5,
}

export interface MonsterStats {
  hp: number;
  attack: number;
  defense: number;
  spAttack: number;
  spDefense: number;
  speed: number;
}
// ...
export enum StatusEffect {
  None = 0,
  Burn = 1,
  Freeze = 2,
  Paralysis = 3,
  Poison = 4,
  BadPoison = 5,
  Sleep = 6,
}
// ...
export interface MoveInstance {
  moveId: number;
  pp: number;
  maxPp: number;
}

export interface PokemonInstance {
  id: string; // Unique ID
  speciesId: number;
  formId?: string;
  level: number;
  experience: number;
  nature: Nature;
  ability: string;
  ivs: MonsterStats;
  evs: MonsterStats;
  stats: MonsterStats;
  currentHp: number;
  friendship: number;
  heldItemId?: number;
  moves: MoveInstance[];
  shiny: boolean;
  otId: string;
  otName: string;
  nickname?: string;
  status?: StatusEffect;
}